# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/confirm": {
            "post": {
                "description": "Confirm user email with verification token or OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Confirm email",
                "parameters": [
                    {
                        "description": "Confirmation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ConfirmEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email confirmed successfully",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/auth/confirm-token": {
            "post": {
                "description": "Confirm user email using Supabase access token from callback URL fragment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Confirm email with access token",
                "parameters": [
                    {
                        "description": "Access token from Supabase",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ConfirmEmailWithTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email confirmed successfully",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
...
        "/auth/resend-confirmation": {
            "post": {
                "description": "Resend confirmation email to user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Resend confirmation email",
                "parameters": [
                    {
                        "description": "Email address",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ResendConfirmationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmation email sent",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/billing/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Initiate a Stripe Checkout session for upgrading to a paid plan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Create Stripe checkout session",
                "parameters": [
                    {
                        "description": "Checkout request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/billing.CheckoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stripe session URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/billing/portal": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a Stripe Customer Portal session to manage subscription, payment methods, and invoices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Create Stripe billing portal session",
                "parameters": [
                    {
                        "description": "Portal request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/billing.PortalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stripe portal URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "404": {
                        "description": "No subscription found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/billing/subscription": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the authenticated user's active subscription and plan details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Get user subscription",
                "responses": {
                    "200": {
                        "description": "Current subscription",
                        "schema": {
                            "$ref": "#/definitions/billing.SubscriptionResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "404": {
                        "description": "No active subscription",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/billing/usage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the authenticated user's resource usage for the current billing period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Get usage stats",
                "responses": {
                    "200": {
                        "description": "Current period usage",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/billing.UsageResult"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/billing/webhook": {
            "post": {
                "description": "Handle Stripe webhook events for subscription lifecycle management",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Stripe webhook handler",
                "responses": {
                    "200": {
                        "description": "Webhook processed"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/domains": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all domains belonging to the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Get user domains",
                "responses": {
                    "200": {
                        "description": "List of user domains",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domains.DomainResult"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new domain for the authenticated user with user-provided encryption public key and optional webhook configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Create a new domain",
                "parameters": [
                    {
                        "description": "Domain creation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domains.CreateDomainRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Domain created successfully",
                        "schema": {
                            "$ref": "#/definitions/domains.DomainResult"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "402": {
                        "description": "Plan limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/domains.planLimitExceededResponse"
                        }
                    }
                }
            }
        },
...
        "/domains/{domainId}/providers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all email providers configured for a specific domain owned by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "List email providers for a domain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Domain ID",
                        "name": "domainId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of providers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/providers.EmailProviderResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden - domain not owned by user",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Domain not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/auth.UserResult"
                }
            }
        },
        "auth.ConfirmEmailRequest": {
            "type": "object",
            "required": [
                "email",
                "token"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "auth.ConfirmEmailWithTokenRequest": {
            "type": "object",
            "required": [
                "access_token"
            ],
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
...
        "auth.ResendConfirmationRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "billing.CheckoutRequest": {
            "type": "object",
            "properties": {
                "cancel_url": {
                    "type": "string"
                },
                "plan_id": {
                    "type": "string"
                },
                "success_url": {
                    "type": "string"
                }
            }
        },
...
        "billing.PortalRequest": {
            "type": "object",
            "properties": {
                "return_url": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponseBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
    }
}